requirements.txt
google-generativeai==0.5.4
python-telegram-bot==21.1.1
python-dotenv==1.0.1
